Final edit to code

+++ signal_handling.c
@@ -5,7 +5,10 @@ void ctrl_c_execution(int signum)
     int ppid = getpid();
     if(ppid == shellpid)
     {
-        kill(fprocpid, SIGINT);
+        if(fprocpid != -1)
+        {
+            kill(fprocpid, SIGINT);
+        }
     }
     else
     {
@@ -19,40 +22,43 @@ void ctrl_z_execution(int signum)
     int ppid = getpid();
     if(ppid == shellpid)
     {
-        kill(fprocpid, SIGTTIN);
-        kill(fprocpid, SIGTSTP);
-        struct back *newback = malloc(sizeof(struct back));
-        struct back *prev = NULL;
-        struct back *curr = present;
-        backnum++;
-        strcpy(newback->backname, fproc->fname);
-        newback->backpid = fprocpid;
-        newback->jobnum = backnum;
-        if(curr == NULL)
+        if(fprocpid != -1)
         {
-            newback->next = NULL;
-            present = newback;
-        }
-        else
-        {
-            while(strcmp(fproc->fname, curr->backname) >= 0)
+            kill(fprocpid, SIGTTIN);
+            kill(fprocpid, SIGTSTP);
+            struct back *newback = malloc(sizeof(struct back));
+            struct back *prev = NULL;
+            struct back *curr = present;
+            backnum++;
+            strcpy(newback->backname, fproc->fname);
+            newback->backpid = fprocpid;
+            newback->jobnum = backnum;
+            if(curr == NULL)
             {
-                prev = curr;
-                curr = curr->next;
-                if(curr == NULL)
-                {
-                    break;
-                }
-            }
-            if(prev != NULL)
-            {
-                prev->next = newback;
-                newback->next = curr;
+                newback->next = NULL;
+                present = newback;
             }
             else
             {
-                newback->next = curr;
-                present = newback;
+                while(strcmp(fproc->fname, curr->backname) >= 0)
+                {
+                    prev = curr;
+                    curr = curr->next;
+                    if(curr == NULL)
+                    {
+                        break;
+                    }
+                }
+                if(prev != NULL)
+                {
+                    prev->next = newback;
+                    newback->next = curr;
+                }
+                else
+                {
+                    newback->next = curr;
+                    present = newback;
+                }
             }
         }
     }
